pyproject.toml
[project]
name = "tts-backend"
version = "0.1.0"
description = "FastAPI backend for TTS service using Alibaba Cloud DashScope"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0"
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"